pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "pip>=21.3"]
build-backend = "setuptools.build_meta"


[project]
name = "culicidaelab"
version = "0.2.2"
description = "A Python library for mosquito detection, segmentation, and classification in images"
readme = "README.md"
authors = [
    { name = "Ilona Kovaleva", email = "iloncka.ds@gmail.com" }
]
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Image Recognition",
]
dependencies = [
    "torch>=2.3.1",
    "torchvision>=0.18.1",
    "numpy>=1.24.4",
    "tqdm>=4.66.1",
    "iopath>=0.1.10",
    "pillow>=9.4.0",
    "fastai>=2.7.0, <=2.8.0",
    "timm>=0.9.0",
    "ultralytics>=8.3.0",
    "pyyaml>=6.0.0",
    "huggingface-hub>=0.16.0",
    "datasets==4.0.0", # >=2.13.0,<=3.1.0
    "omegaconf>=2.3.0",
    "python-dotenv>=0.21.0",
    "pycocotools>=2.0.7",
    "dill>=0.3.8",
    "graphviz>=0.20.3",
    "appdirs>=1.4.4",
    "requests>=2.25.1",
    "toml>=0.10.2",
    "dotenv>=0.9.9",
    "pydantic-settings>=2.9.1",
    "onnxruntime>=1.22.1",
]


[project.urls]
"Homepage" = "https://github.com/iloncka/culicidaelab"
"Bug Tracker" = "https://github.com/iloncka/culicidaelab/issues"


[tool.setuptools.packages.find]
exclude = [
    "tests*",
    "example*",
    "docs*",
    "scripts*",
    "site*"
]

[tool.setuptools.package-data]
"*" = ["**/*.yaml"]


[project.optional-dependencies]
examples = [
    "ipykernel==6.16.2",
    "ipywidgets>=8.1.7",
    "jupytext>=1.17.2",
]
docs = [
    "culicidaelab[examples]",
    "mkdocs[i18n]>=1.6.1",
    "mkdocs-material==9.5.46",
    "mkdocstrings[python]==0.27.0",
    "mkdocs-custom-fences>=0.1.2",
    "mkdocs-gallery>=0.10.4",
]
test = [
    "black==23.3.0",
    "mypy==1.4.1",
    "pre-commit==2.21.0",
    "pytest==7.4.4",
    "pytest-cov==4.1.0",
    "pytest-mock==3.14.1",
    "ruff==0.8.0",
    "twine==5.1.1",
    "bandit==1.8.5",
    "types-requests==2.32.4.20250611",
    "types-pyyaml==6.0.12.20250516",
    "types-toml==0.10.8.20240310",
    "gputil==1.4.0",
    "requests-mock>=1.12.1",
    "tabulate>=0.9.0",
]
dev = [
    "culicidaelab[docs]",
    "culicidaelab[test]"
]
